refactor(canon): move new tests out of over-limit files
The source-file-lengths check fails when already over-limit files grow.
Relocate the two regression tests added in this PR to files that are
within the 350-line limit so the flagged files do not grow:

- lsp_client/session.rs: move keeps_encoded_vue_virtual_overlay_uri_at_real_path_inside_project to lsp_client/tests.rs (which already holds the lsp_client integration tests)
- corsa_bridge/bridge.rs: move normalizes_absolute_paths_with_file_uri_encoding to corsa_bridge.rs test block; expose normalize_document_uri as pub(super) and inline the path_to_file_uri call to avoid adding import lines
